vsock: separate the implementations per-platform
Previously when we supported the non-upstream 4.9 patches we had to
reference `hvsock` even when on Linux.

Now that we don't support these non-upstream patches we can move the
`hvsock` support to Windows and `vsock` to Linux.

This fixes cross-compilation, specifically this error:

$ GOOS=linux go build
../../../vendor/github.com/linuxkit/virtsock/pkg/hvsock/hvsock.go:158:10: undefined: hvsockListener
../../../vendor/github.com/linuxkit/virtsock/pkg/hvsock/hvsock.go:173:10: undefined: hvsockListener
../../../vendor/github.com/linuxkit/virtsock/pkg/hvsock/hvsock.go:178:10: undefined: hvsockListener
../../../vendor/github.com/linuxkit/virtsock/pkg/hvsock/hvsock.go:205:2: undefined: hvsockConn
