+++ go/pkg/vpnkit/transport/vsock_windows_test.go
@@ -5,7 +5,6 @@ import (
 	"testing"
 
 	"github.com/linuxkit/virtsock/pkg/hvsock"
-	"github.com/linuxkit/virtsock/pkg/vsock"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -13,8 +12,6 @@ func TestParse(t *testing.T) {
 	a, err := parseAddr("2")
 	assert.Nil(t, err)
 	assert.EqualValues(t, hvsock.GUIDZero, a.vmID)
-	assert.EqualValues(t, vsock.CIDAny, a.cid)
-	assert.Equal(t, uint32(2), a.port)
 	assert.Equal(t, fmt.Sprintf("%08x-facb-11e6-bd58-64006a7986d3", 2), a.svcID.String())
 }
 
@@ -22,7 +19,6 @@ func TestParseSvcID(t *testing.T) {
 	a, err := parseAddr(fmt.Sprintf("%08x-facb-11e6-bd58-64006a7986d3", 2))
 	assert.Nil(t, err)
 	assert.EqualValues(t, hvsock.GUIDZero, a.vmID)
-	assert.EqualValues(t, vsock.CIDAny, a.cid)
 	// we don't try to compute the corresponding AF_VSOCK
 	assert.Equal(t, fmt.Sprintf("%08x-facb-11e6-bd58-64006a7986d3", 2), a.svcID.String())
 }
@@ -32,7 +28,6 @@ func TestParseVmIDPort(t *testing.T) {
 	a, err := parseAddr(fmt.Sprintf("%s/%08x-facb-11e6-bd58-64006a7986d3", vmID, 2))
 	assert.Nil(t, err)
 	assert.EqualValues(t, vmID, a.vmID.String())
-	assert.EqualValues(t, vsock.CIDAny, a.cid)
 	assert.Equal(t, fmt.Sprintf("%08x-facb-11e6-bd58-64006a7986d3", 2), a.svcID.String())
 }
 
@@ -41,8 +36,6 @@ func TestParseVmIDPort2(t *testing.T) {
 	a, err := parseAddr(fmt.Sprintf("%s/%d", vmID, 2))
 	assert.Nil(t, err)
 	assert.EqualValues(t, vmID, a.vmID.String())
-	assert.EqualValues(t, vsock.CIDAny, a.cid)
-	assert.Equal(t, uint32(2), a.port)
 	assert.Equal(t, fmt.Sprintf("%08x-facb-11e6-bd58-64006a7986d3", 2), a.svcID.String())
 }
 
@@ -50,7 +43,5 @@ func TestParseCIDPort(t *testing.T) {
 	a, err := parseAddr("3/2")
 	assert.Nil(t, err)
 	assert.EqualValues(t, hvsock.GUIDZero, a.vmID)
-	assert.EqualValues(t, uint32(3), a.cid)
-	assert.Equal(t, uint32(2), a.port)
 	assert.Equal(t, fmt.Sprintf("%08x-facb-11e6-bd58-64006a7986d3", 2), a.svcID.String())
 }
